# Conflicts:
#	interactions/client.py

+++ interactions/api/models/channel.py
@@ -1,8 +1,8 @@
 from datetime import datetime
 from enum import IntEnum
-from typing import Optional
+from typing import List, Optional, Union
 
-from .misc import DictSerializerMixin, Snowflake
+from .misc import MISSING, DictSerializerMixin, Snowflake
 
 
 class ChannelType(IntEnum):
@@ -159,6 +159,7 @@ class Channel(DictSerializerMixin):
         "_client",
         # TODO: Document banner when Discord officially documents them.
         "banner",
+        "guild_hashes",
     )
 
     def __init__(self, **kwargs):
@@ -182,16 +183,25 @@ class Channel(DictSerializerMixin):
 
     async def send(
         self,
-        content: Optional[str] = None,
+        content: Optional[str] = MISSING,
         *,
-        tts: Optional[bool] = False,
+        tts: Optional[bool] = MISSING,
         # attachments: Optional[List[Any]] = None,  # TODO: post-v4: Replace with own file type.
-        embeds=None,
-        allowed_mentions=None,
-        components=None,
+        embeds: Optional[Union["Embed", List["Embed"]]] = MISSING,  # noqa
+        allowed_mentions: Optional["MessageInteraction"] = MISSING,  # noqa
+        components: Optional[
+            Union[
+                "ActionRow",  # noqa
+                "Button",  # noqa
+                "SelectMenu",  # noqa
+                List["ActionRow"],  # noqa
+                List["Button"],  # noqa
+                List["SelectMenu"],  # noqa
+            ]
+        ] = MISSING,
     ):
         """
-        Sends a message in the channel
+        Sends a message in the channel.
 
         :param content?: The contents of the message as a string or string-converted value.
         :type content: Optional[str]
@@ -202,35 +212,127 @@ class Channel(DictSerializerMixin):
         :param allowed_mentions?: The message interactions/mention limits that the message can refer to.
         :type allowed_mentions: Optional[MessageInteraction]
         :param components?: A component, or list of components for the message.
-        :type components: Optional[Union[Component, List[Component]]]
+        :type components: Optional[Union[ActionRow, Button, SelectMenu, List[Actionrow], List[Button], List[SelectMenu]]]
         :return: The sent message as an object.
         :rtype: Message
         """
+        if not self._client:
+            raise AttributeError("HTTPClient not found!")
         from ...models.component import ActionRow, Button, SelectMenu
         from .message import Message
 
-        _content: str = "" if content is None else content
-        _tts: bool = False if tts is None else tts
+        _content: str = "" if content is MISSING else content
+        _tts: bool = False if tts is MISSING else tts
         # _file = None if file is None else file
         # _attachments = [] if attachments else None
-        _embeds: list = []
-        _allowed_mentions: dict = {} if allowed_mentions is None else allowed_mentions
-        _components: list = [{"type": 1, "components": []}]
-        if embeds:
+        _allowed_mentions: dict = {} if allowed_mentions is MISSING else allowed_mentions
+        if not embeds or embeds is MISSING:
+            _embeds: list = []
+        else:
             if isinstance(embeds, list):
                 _embeds = [embed._json for embed in embeds]
             else:
                 _embeds = [embeds._json]
 
-        if isinstance(components, ActionRow):
-            _components[0]["components"] = [component._json for component in components.components]
-        elif isinstance(components, Button):
-            _components[0]["components"] = [] if components is None else [components._json]
-        elif isinstance(components, SelectMenu):
-            components._json["options"] = [option._json for option in components.options]
-            _components[0]["components"] = [] if components is None else [components._json]
+        # TODO: Break this obfuscation pattern down to a "builder" method.
+        if not components or components is MISSING:
+            _components = []
         else:
-            _components = [] if components is None else [components]
+            _components: List[dict] = [{"type": 1, "components": []}]
+            if isinstance(components, list) and all(
+                isinstance(action_row, ActionRow) for action_row in components
+            ):
+                _components = [
+                    {
+                        "type": 1,
+                        "components": [
+                            (
+                                component._json
+                                if component._json.get("custom_id") or component._json.get("url")
+                                else []
+                            )
+                            for component in action_row.components
+                        ],
+                    }
+                    for action_row in components
+                ]
+            elif isinstance(components, list) and all(
+                isinstance(component, (Button, SelectMenu)) for component in components
+            ):
+                for component in components:
+                    if isinstance(component, SelectMenu):
+                        component._json["options"] = [
+                            options._json if not isinstance(options, dict) else options
+                            for options in component._json["options"]
+                        ]
+                _components = [
+                    {
+                        "type": 1,
+                        "components": [
+                            (
+                                component._json
+                                if component._json.get("custom_id") or component._json.get("url")
+                                else []
+                            )
+                            for component in components
+                        ],
+                    }
+                ]
+            elif isinstance(components, list) and all(
+                isinstance(action_row, (list, ActionRow)) for action_row in components
+            ):
+                _components = []
+                for action_row in components:
+                    for component in (
+                        action_row if isinstance(action_row, list) else action_row.components
+                    ):
+                        if isinstance(component, SelectMenu):
+                            component._json["options"] = [
+                                option._json for option in component.options
+                            ]
+                    _components.append(
+                        {
+                            "type": 1,
+                            "components": [
+                                (
+                                    component._json
+                                    if component._json.get("custom_id")
+                                    or component._json.get("url")
+                                    else []
+                                )
+                                for component in (
+                                    action_row
+                                    if isinstance(action_row, list)
+                                    else action_row.components
+                                )
+                            ],
+                        }
+                    )
+            elif isinstance(components, ActionRow):
+                _components[0]["components"] = [
+                    (
+                        component._json
+                        if component._json.get("custom_id") or component._json.get("url")
+                        else []
+                    )
+                    for component in components.components
+                ]
+            elif isinstance(components, Button):
+                _components[0]["components"] = (
+                    [components._json]
+                    if components._json.get("custom_id") or components._json.get("url")
+                    else []
+                )
+            elif isinstance(components, SelectMenu):
+                components._json["options"] = [
+                    options._json if not isinstance(options, dict) else options
+                    for options in components._json["options"]
+                ]
+                _components[0]["components"] = (
+                    [components._json]
+                    if components._json.get("custom_id") or components._json.get("url")
+                    else []
+                )
 
         # TODO: post-v4: Add attachments into Message obj.
         payload = Message(
@@ -250,23 +352,26 @@ class Channel(DictSerializerMixin):
         """
         Deletes the channel.
         """
+        if not self._client:
+            raise AttributeError("HTTPClient not found!")
         await self._client.delete_channel(channel_id=int(self.id))
 
     async def modify(
         self,
-        name: Optional[str] = None,
-        topic: Optional[str] = None,
-        bitrate: Optional[int] = None,
-        user_limit: Optional[int] = None,
-        rate_limit_per_user: Optional[int] = None,
-        position: Optional[int] = None,
+        name: Optional[str] = MISSING,
+        topic: Optional[str] = MISSING,
+        bitrate: Optional[int] = MISSING,
+        user_limit: Optional[int] = MISSING,
+        rate_limit_per_user: Optional[int] = MISSING,
+        position: Optional[int] = MISSING,
         # permission_overwrites,
-        parent_id: Optional[int] = None,
-        nsfw: Optional[bool] = False,
+        parent_id: Optional[int] = MISSING,
+        nsfw: Optional[bool] = MISSING,
         reason: Optional[str] = None,
     ) -> "Channel":
         """
-        Edits the channel
+        Edits the channel.
+
         :param name?: The name of the channel, defaults to the current value of the channel
         :type name: str
         :param topic?: The topic of that channel, defaults to the current value of the channel
@@ -288,16 +393,18 @@ class Channel(DictSerializerMixin):
         :return: The modified channel as new object
         :rtype: Channel
         """
-        _name = self.name if not name else name
-        _topic = self.topic if not topic else topic
-        _bitrate = self.bitrate if not bitrate else bitrate
-        _user_limit = self.user_limit if not user_limit else user_limit
+        if not self._client:
+            raise AttributeError("HTTPClient not found!")
+        _name = self.name if name is MISSING else name
+        _topic = self.topic if topic is MISSING else topic
+        _bitrate = self.bitrate if bitrate is MISSING else bitrate
+        _user_limit = self.user_limit if user_limit is MISSING else user_limit
         _rate_limit_per_user = (
-            self.rate_limit_per_user if not rate_limit_per_user else rate_limit_per_user
+            self.rate_limit_per_user if rate_limit_per_user is MISSING else rate_limit_per_user
         )
-        _position = self.position if not position else position
-        _parent_id = self.parent_id if not parent_id else parent_id
-        _nsfw = self.nsfw if not nsfw else nsfw
+        _position = self.position if position is MISSING else position
+        _parent_id = self.parent_id if parent_id is MISSING else parent_id
+        _nsfw = self.nsfw if nsfw is MISSING else nsfw
         _type = self.type
 
         payload = Channel(
@@ -323,16 +430,83 @@ class Channel(DictSerializerMixin):
         member_id: int,
     ) -> None:
         """
-        This adds a member to the channel, if the channel is a thread
+        This adds a member to the channel, if the channel is a thread.
 
         :param member_id: The id of the member to add to the channel
         :type member_id: int
         """
+        if not self._client:
+            raise AttributeError("HTTPClient not found!")
         if not self.thread_metadata:
             raise TypeError(
                 "The Channel you specified is not a thread!"
             )  # TODO: Move to new error formatter.
         await self._client.add_member_to_thread(thread_id=int(self.id), user_id=member_id)
+
+    async def pin_message(
+        self,
+        message_id: int,
+    ) -> None:
+        """
+        Pins a message to the channel.
+
+        :param message_id: The id of the message to pin
+        :type message_id: int
+        """
+        if not self._client:
+            raise AttributeError("HTTPClient not found!")
+
+        await self._client.pin_message(channel_id=int(self.id), message_id=message_id)
+
+    async def unpin_message(
+        self,
+        message_id: int,
+    ) -> None:
+        """
+        Unpins a message from the channel.
+
+        :param message_id: The id of the message to unpin
+        :type message_id: int
+        """
+        if not self._client:
+            raise AttributeError("HTTPClient not found!")
+
+        await self._client.unpin_message(channel_id=int(self.id), message_id=message_id)
+
+    async def publish_message(
+        self,
+        message_id: int,
+    ):
+        """Publishes (API calls it crossposts) a message in the channel to any that is followed by.
+
+        :param message_id: The id of the message to publish
+        :type message_id: int
+        :return: The message published
+        :rtype: Message
+        """
+        if not self._client:
+            raise AttributeError("HTTPClient not found!")
+        from .message import Message
+
+        res = await self._client.publish_message(
+            channel_id=int(self.id), message_id=int(message_id)
+        )
+        return Message(**res, _client=self._client)
+
+    async def get_pinned_messages(self):
+        """
+        Get all pinned messages from the channel.
+
+        :return: A list of pinned message objects.
+        :rtype: List[Message]
+        """
+        if not self._client:
+            raise AttributeError("HTTPClient not found!")
+        from .message import Message
+
+        res = await self._client.get_pinned_messages(int(self.id))
+        messages = [Message(**message, _client=self._client) for message in res]
+        return messages
 
 
 class Thread(Channel):
