+++ interactions/api/models/role.py
@@ -1,6 +1,6 @@
-from typing import Optional
+from typing import List, Optional
 
-from .misc import DictSerializerMixin, Snowflake
+from .misc import MISSING, DictSerializerMixin, Snowflake
 
 
 class RoleTags(DictSerializerMixin):
@@ -69,13 +69,15 @@ class Role(DictSerializerMixin):
         reason: Optional[str] = None,
     ) -> None:
         """
-        Deletes the role from the guild
+        Deletes the role from the guild.
+
         :param guild_id: The id of the guild to delete the role from
         :type guild_id: int
         :param reason: The reason for the deletion
         :type reason: Optional[str]
         """
-
+        if not self._client:
+            raise AttributeError("HTTPClient not found!")
         await self._client.delete_guild_role(
             guild_id=guild_id, role_id=int(self.id), reason=reason
         ),
@@ -83,17 +85,18 @@ class Role(DictSerializerMixin):
     async def modify(
         self,
         guild_id: int,
-        name: Optional[str] = None,
+        name: Optional[str] = MISSING,
         # permissions,
-        color: Optional[int] = None,
-        hoist: Optional[bool] = None,
+        color: Optional[int] = MISSING,
+        hoist: Optional[bool] = MISSING,
         # icon,
         # unicode_emoji,
-        mentionable: Optional[bool] = None,
+        mentionable: Optional[bool] = MISSING,
         reason: Optional[str] = None,
     ) -> "Role":
         """
-        Edits the role in a guild
+        Edits the role in a guild.
+
         :param guild_id: The id of the guild to edit the role on
         :type guild_id: int
         :param name?: The name of the role, defaults to the current value of the role
@@ -109,11 +112,12 @@ class Role(DictSerializerMixin):
         :return: The modified role object
         :rtype: Role
         """
-
-        _name = self.name if not name else name
-        _color = self.color if not color else color
-        _hoist = self.hoist if not hoist else hoist
-        _mentionable = self.mentionable if mentionable is None else mentionable
+        if not self._client:
+            raise AttributeError("HTTPClient not found!")
+        _name = self.name if name is MISSING else name
+        _color = self.color if color is MISSING else color
+        _hoist = self.hoist if hoist is MISSING else hoist
+        _mentionable = self.mentionable if mentionable is MISSING else mentionable
 
         payload = Role(name=_name, color=_color, hoist=_hoist, mentionable=_mentionable)
 
@@ -124,3 +128,29 @@ class Role(DictSerializerMixin):
             reason=reason,
         )
         return Role(**res, _client=self._client)
+
+    async def modify_position(
+        self,
+        guild_id: int,
+        position: int,
+        reason: Optional[str] = None,
+    ) -> List["Role"]:
+        """
+        Modifies the position of a role in the guild.
+
+        :param guild_id: The id of the guild to modify the role position on
+        :type guild_id: int
+        :param position: The new position of the role
+        :type position: int
+        :param reason?: The reason for the modifying
+        :type reason: Optional[str]
+        :return: List of guild roles with updated hierarchy
+        :rtype: List[Role]
+        """
+        if not self._client:
+            raise AttributeError("HTTPClient not found!")
+        res = await self._client.modify_guild_role_position(
+            guild_id=guild_id, position=position, role_id=int(self.id), reason=reason
+        )
+        roles = [Role(**role, _client=self._client) for role in res]
+        return roles
